AS-2023-2024/5X-TPSIT/auth+jwt+sqlite/requirements.txt
httptools
uvicorn
fastapi
pydantic
jinja2
sqlalchemy
pymysql
pyjwt
python-decouple